Refactor: exception 패키지를 community 패키지 하위로 옮김
resolves: #82, #164
