+++ src/main/java/com/filmdoms/community/account/data/dto/request/UpdateProfileRequestDto.java
@@ -1,8 +1,8 @@
 package com.filmdoms.community.account.data.dto.request;
 
-import static com.filmdoms.community.account.exception.ValidationMessage.IMAGE_REQUIRED;
-import static com.filmdoms.community.account.exception.ValidationMessage.LIST_TOO_BIG;
-import static com.filmdoms.community.account.exception.ValidationMessage.UNMATCHED_NICKNAME;
+import static com.filmdoms.community.exception.ValidationMessage.IMAGE_REQUIRED;
+import static com.filmdoms.community.exception.ValidationMessage.LIST_TOO_BIG;
+import static com.filmdoms.community.exception.ValidationMessage.UNMATCHED_NICKNAME;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
