+++ src/main/java/com/filmdoms/community/account/data/dto/request/profile/UpdateProfileImageRequestDto.java
@@ -3,7 +3,7 @@ package com.filmdoms.community.account.data.dto.request.profile;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
-import static com.filmdoms.community.account.exception.ValidationMessage.IMAGE_REQUIRED;
+import static com.filmdoms.community.exception.ValidationMessage.IMAGE_REQUIRED;
 
 @Getter
 public class UpdateProfileImageRequestDto {
