+++ src/main/java/com/filmdoms/community/account/service/EmailService.java
@@ -3,8 +3,8 @@ package com.filmdoms.community.account.service;
 import com.filmdoms.community.account.data.dto.request.AuthCodeVerificationRequestDto;
 import com.filmdoms.community.account.data.dto.response.EmailAuthDto;
 import com.filmdoms.community.account.data.entity.Account;
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import com.filmdoms.community.account.repository.AccountRepository;
 import com.filmdoms.community.account.service.utils.PasswordUtil;
 import com.filmdoms.community.account.service.utils.RedisUtil;
