+++ src/main/java/com/filmdoms/community/article/controller/AnnounceController.java
@@ -3,8 +3,8 @@ package com.filmdoms.community.article.controller;
 import com.filmdoms.community.account.data.constant.AccountRole;
 import com.filmdoms.community.account.data.dto.AccountDto;
 import com.filmdoms.community.account.data.dto.response.Response;
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import com.filmdoms.community.article.data.constant.Category;
 import com.filmdoms.community.article.data.dto.request.announce.AnnounceRegisterDto;
 import com.filmdoms.community.article.data.dto.response.boardlist.AnnounceListResponseDto;
