+++ src/main/java/com/filmdoms/community/article/data/constant/Tag.java
@@ -1,7 +1,7 @@
 package com.filmdoms.community.article.data.constant;
 
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
