+++ src/main/java/com/filmdoms/community/article/data/dto/request/create/ParentCreateRequestDto.java
@@ -12,7 +12,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import static com.filmdoms.community.account.exception.ValidationMessage.*;
+import static com.filmdoms.community.exception.ValidationMessage.*;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME,
         include = JsonTypeInfo.As.EXISTING_PROPERTY,
