+++ src/main/java/com/filmdoms/community/board/post/service/PostService.java
@@ -1,8 +1,8 @@
 package com.filmdoms.community.board.post.service;
 
 import com.filmdoms.community.account.data.dto.AccountDto;
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import com.filmdoms.community.account.repository.AccountRepository;
 import com.filmdoms.community.board.data.BoardContent;
 import com.filmdoms.community.board.post.data.dto.PostBriefDto;
