+++ src/main/java/com/filmdoms/community/config/SecurityConfig.java
@@ -5,8 +5,8 @@ import com.filmdoms.community.config.jwt.JwtTokenProvider;
 import com.filmdoms.community.config.oauth.CustomOAuth2AuthorizationRequestResolver;
 import com.filmdoms.community.config.oauth.CustomOAuthFailureHandler;
 import com.filmdoms.community.config.oauth.CustomOAuthSuccessHandler;
-import com.filmdoms.community.account.exception.CustomAccessDeniedHandler;
-import com.filmdoms.community.account.exception.CustomAuthenticationEntryPoint;
+import com.filmdoms.community.exception.CustomAccessDeniedHandler;
+import com.filmdoms.community.exception.CustomAuthenticationEntryPoint;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
