+++ src/main/java/com/filmdoms/community/config/oauth/CustomOAuthFailureHandler.java
@@ -3,8 +3,8 @@ package com.filmdoms.community.config.oauth;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.filmdoms.community.config.jwt.JwtTokenProvider;
 import com.filmdoms.community.account.data.dto.response.Response;
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
