+++ src/main/java/com/filmdoms/community/config/oauth/CustomOAuthSuccessHandler.java
@@ -7,8 +7,8 @@ import com.filmdoms.community.account.data.constant.OAuthType;
 import com.filmdoms.community.account.data.dto.response.OAuthResponseDto;
 import com.filmdoms.community.account.data.dto.response.Response;
 import com.filmdoms.community.account.data.entity.Account;
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import com.filmdoms.community.account.repository.AccountRepository;
 import com.filmdoms.community.account.repository.RefreshTokenRepository;
 import jakarta.servlet.http.HttpServletRequest;
