+++ src/test/java/com/filmdoms/community/account/controller/AccountControllerTest.java
@@ -6,8 +6,8 @@ import com.filmdoms.community.config.jwt.JwtTokenProvider;
 import com.filmdoms.community.account.data.dto.LoginDto;
 import com.filmdoms.community.account.data.dto.request.LoginRequestDto;
 import com.filmdoms.community.account.data.dto.response.AccessTokenResponseDto;
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import com.filmdoms.community.account.repository.AccountRepository;
 import com.filmdoms.community.account.service.AccountService;
 import java.util.List;
