+++ src/test/java/com/filmdoms/community/banner/service/BannerServiceTest.java
@@ -3,8 +3,8 @@ package com.filmdoms.community.banner.service;
 import com.filmdoms.community.account.data.constant.AccountRole;
 import com.filmdoms.community.account.data.dto.AccountDto;
 import com.filmdoms.community.account.data.entity.Account;
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import com.filmdoms.community.account.repository.AccountRepository;
 import com.filmdoms.community.banner.data.dto.request.BannerRequestDto;
 import com.filmdoms.community.banner.data.dto.response.BannerResponseDto;
