+++ src/test/java/com/filmdoms/community/vote/controller/VoteControllerTest.java
@@ -1,7 +1,7 @@
 package com.filmdoms.community.vote.controller;
 
-import com.filmdoms.community.account.exception.ApplicationException;
-import com.filmdoms.community.account.exception.ErrorCode;
+import com.filmdoms.community.exception.ApplicationException;
+import com.filmdoms.community.exception.ErrorCode;
 import com.filmdoms.community.testconfig.TestSecurityConfig;
 import com.filmdoms.community.vote.data.dto.VoteResponseDto;
 import com.filmdoms.community.vote.service.VoteService;
